# swagger.yaml
...
swagger: '2.0'
info:
  title: IoT Device Repository
  termsOfService: http://swagger.io/terms/
  description: 'The IOT Device Repository provides metadata for all IOT devices managed on the platform, their device classes and the associated data structures for the user.'
  version: '1.0'
  contact:
    name: sh20xyqi
    email: sh20xyqi@studserv.uni-leipzig.de
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
schemes:
  - https
host: api.sepl.infai.org
basePath: /iot-device-repo
produces:
  - text/plain
...
paths:
...
  /other/vendor:
    post:
      tags:
        - otherStructs
      summary:
        Adds a new vendor.
      description:
        ''
      produces:
        - application/json
      parameters:
      - in: body
        name: body
        description: Vendor object that needs to be added to the database.
        required: true
        schema:
          $ref: '#/definitions/Vendor'
      responses:
        200:
          description: Returns the instanciated vendor Id.
          schema:
            $ref: '#/definitions/Insert_OK'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - JSON encoding failed; Vendor instance can not created in database.
          schema:
            $ref: '#/definitions/ErrorMessage'
  /other/protocol:
    post:
      tags:
        - otherStructs
      summary:
        Adds a new protocol.
      description:
        ''
      produces:
        - application/json
      parameters:
      - in: body
        name: body
        description: Protocol object that needs to be added to the database.
        required: true
        schema:
          $ref: '#/definitions/Protocol'
      responses:
        200:
          description: Returns the instanciated protocol Id.
          schema:
            $ref: '#/definitions/Insert_OK'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - JSON encoding failed; Protocol instance can not created in database.
          schema:
            $ref: '#/definitions/ErrorMessage'
  /other/deviceclass:
    post:
      tags:
        - otherStructs
      summary:
        Adds a new device class.
      description:
        ''
      produces:
        - application/json
      parameters:
      - in: body
        name: body
        description: Device class object that needs to be added to the database.
        required: true
        schema:
          $ref: '#/definitions/DeviceClass'
      responses:
        200:
          description: Returns the instanciated device class Id.
          schema:
            $ref: '#/definitions/Insert_OK'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - JSON encoding failed; Device class instance can not created in database.
          schema:
            $ref: '#/definitions/ErrorMessage'
  /other/valueType:
    post:
      tags:
        - otherStructs
      summary:
        Adds a new value type.
      description:
        ''
      produces:
        - application/json
      parameters:
      - in: body
        name: body
        description: Value type object that needs to be added to the database.
        required: true
        schema:
          $ref: '#/definitions/ValueType'
      responses:
        200:
          description: Returns the instanciated value type Id.
          schema:
            $ref: '#/definitions/Insert_OK'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - JSON encoding failed; Value type instance can not created in database.
          schema:
            $ref: '#/definitions/ErrorMessage'
...
  /ui/search/deviceTypes/{query}/{limit}/{offset}:
    get:
      tags:
        - search
      summary:
        Returns device types that use a specific query, within a limit and how much elements had to be skipped.
      description:
        ''
      produces:
        - application/json
      parameters:
        - name: query
          in: path
          description: Identifies all device types that use this specific query.
          required: true
          type: string
        - name: limit
          in: path
          description: Restricts the set of returnable elements to the passed limit.
          required: true
          type: integer
        - name: offset
          in: path
          description: Specifies how many elements of the set are skipped.
          required: true
          type: integer
      responses:
        200:
          description: Returns all device types matched with the passed query.
          schema:
            type: array
            items:
              $ref: '#/definitions/ValueType'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - A specific limit can not be found; a specific offset can not be found; DeviceTypes can not be found in database.
          schema:
            $ref: '#/definitions/ErrorMessage'
  /ui/search/deviceInstances/{query}/{limit}/{offset}:
    get:
      tags:
        - search
      summary:
        Returns device instances that use a specific query, within a limit and how much elements had to be skipped.
      description:
        ''
      produces:
        - application/json
      parameters:
        - name: query
          in: path
          description: Identifies all device instances that use this specific query.
          required: true
          type: string
        - name: limit
          in: path
          description: Restricts the set of returnable elements to the passed limit.
          required: true
          type: integer
        - name: offset
          in: path
          description: Specifies how many elements of the set are skipped.
          required: true
          type: integer
      responses:
        200:
          description: Returns all device instances matched with the passed query.
          schema:
            type: array
            items:
              $ref: '#/definitions/DeviceInstance'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - A specific limit can not be found; a specific offset can not be found; DeviceInstances can not be found in database.
          schema:
            $ref: '#/definitions/ErrorMessage'
  /ui/search/ValueTypes/{query}/{limit}/{offset}:
    get:
      tags:
        - search
      summary:
        Returns value types that use a specific query, within a limit and how much elements had to be skipped.
      description:
        ''
      produces:
        - application/json
      parameters:
        - name: query
          in: path
          description: Identifies all value types that use this specific query.
          required: true
          type: string
        - name: limit
          in: path
          description: Restricts the set of returnable elements to the passed limit.
          required: true
          type: integer
        - name: offset
          in: path
          description: Specifies how many elements of the set are skipped.
          required: true
          type: integer
      responses:
        200:
          description: Returns all value types matched with the passed query.
          schema:
            type: array
            items:
              $ref: '#/definitions/ValueType'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - A specific limit can not be found; a specific offset can not be found; ValueTypes can not be found in database.
          schema:
            $ref: '#/definitions/ErrorMessage'
  /ui/search/others/{type}/{query}/{limit}/{offset}:
    get:
      tags:
        - search
      summary:
        Returns a selectable type of information structure that use a specific query, within a limit and how much elements had to be skipped.
      description:
        ''
      produces:
        - application/json
      parameters:
        - name: type
          in: path
          description: Identifies the selectable information structure. This can be vendors, deviceClasses, protocols or a new, unknown type.
          required: true
          type: string
        - name: query
          in: path
          description: Identifies all type of information structures / instances that use this specific query.
          required: true
          type: string
        - name: limit
          in: path
          description: Restricts the set of returnable elements to the passed limit.
          required: true
          type: integer
        - name: offset
          in: path
          description: Specifies how many elements of the set are skipped.
          required: true
          type: integer
      responses:
        200:
          description: Returns all elements of the selectable information strucutre matched with the passed query.
          schema:
            $ref: ''
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - A specific limit can not be found; a specific offset can not be found; selectable information structure can not be found in database.
          schema:
            $ref: '#/definitions/ErrorMessage'
#Everything that belongs to definitions
definitions:
  MsgSegment:
    type: object
    properties:
      id:
        type: string
        format: string
      name:
        type: string
        format: string
      constraints:
        type: array
        items:
          type: string
    required:
      - id
      - name
      - constraints
  Protocol:
    type: object
    properties:
      id:
        type: string
        format: string
      protocol_handler_url:
        type: string
        format: string
      name:
        type: string
        format: string
      description:
        type: string
        format: string
      msg_structure:
        type: array
        items:
          $ref: '#/definitions/MsgSegment'
    required:
      - id
      - protocol_handler_url
      - name
      - description
      - msg_structure
...
  DeviceInstance:
    type: object
    properties:
      id:
        type: string
        format: string
      name:
        type: string
        format: string
      device_type:
        type: string
        format: string
      config:
        type: array
        items:
          $ref: '#/definitions/ConfigField'
      uri:
        type: string
        format: string
    required:
      - id
      - name
      - device_type
      - config
      - uri
...
  Vendor:
    type: object
    properties:
      id:
        type: string
        format: string
      name:
        type: string
        format: string
    required:
      - id
      - name
  DeviceClass:
    type: object
    properties:
      id:
        type: string
        format: string
      name:
        type: string
        format: string
    required:
      - id
      - name
  ConfigField:
    type: object
    properties:
      id:
        type: string
        format: string
      name:
        type: string
        format: string
      value:
        type: string
        format: string
    required:
      - id
      - name
      - value
...
  FieldType:
    type: object
    properties:
      id:
        type: string
        format: string
      name:
        type: string
        format: string
      type:
        type: object
        format: '#/definitions/ValueType'
    required:
      - id
      - name
      - type
  ValueType:
    type: object
    properties:
      id:
        type: string
        format: string
      name:
        type: string
        format: string
      description:
        type: string
        format: string
      base_type:
        type: string
        format: string
      fields:
        type: array
        items:
          $ref: '#/definitions/FieldType'
      literal:
        type: string
        format: string
    required:
      - id
      - name
      - description
      - base_type
      - fields
      - literal
...
  ErrorMessage:
    type: object
    properties:
      status_code:
        type: integer
        format: int32
      message:
        type: string
        format: string
      error_code:
        type: string
        format: string
      detail:
        type: array
        items:
          type: string
    required:
      - status_code
      - message
      - error_code
      - detail
  Insert_OK:
    type: object
    properties:
      created_id:
        type: string
        format: string
    required:
      - created_id